+++ src/media_sorter/__init__.py
@@ -1,4 +1,4 @@
 """Media Sorter package."""
 
-__version__ = "1.0.5"
+__version__ = "1.0.6"